refactor(api): move business logic to MessageService and extract events
- MessageService now accepts API entities (CreateMessageRequest) and
  handles conversion to database entities internally
- Event emission and workflow triggering moved from controller to service
- Extract SSE event streaming to separate events.rs module
- Controller (messages.rs) now has minimal routing logic only
