+++ crates/everruns-api/src/main.rs
@@ -5,6 +5,7 @@
 mod agents;
 mod auth;
 mod capabilities;
+mod events;
 mod llm_models;
 mod llm_providers;
 mod messages;
@@ -69,7 +70,7 @@ struct HealthState {
         sessions::delete_session,
         messages::create_message,
         messages::list_messages,
-        messages::stream_events,
+        events::stream_events,
         llm_providers::create_provider,
         llm_providers::list_providers,
         llm_providers::get_provider,
@@ -201,6 +202,7 @@ async fn main() -> Result<()> {
     let agents_state = agents::AppState::new(db.clone());
     let sessions_state = sessions::AppState::new(db.clone());
     let messages_state = messages::AppState::new(db.clone(), runner.clone());
+    let events_state = events::AppState::new(db.clone());
     let llm_providers_state = llm_providers::AppState {
         db: db.clone(),
         encryption: encryption.clone(),
@@ -245,6 +247,7 @@ async fn main() -> Result<()> {
         .merge(agents::routes(agents_state))
         .merge(sessions::routes(sessions_state))
         .merge(messages::routes(messages_state))
+        .merge(events::routes(events_state))
         .merge(llm_models::routes(llm_models_state))
         .merge(llm_providers::routes(llm_providers_state))
         .merge(capabilities::routes(capabilities_state))
