+++ crates/everruns-api/src/services/event.rs
@@ -1,5 +1,10 @@
 // Event service for business logic (M2)
 // Events are SSE notifications, NOT primary data storage
+//
+// Note: Currently unused as event emission is handled directly in MessageService.
+// Kept for future use when event emission becomes more complex.
+
+#![allow(dead_code)]
 
 use anyhow::Result;
 use everruns_contracts::Event;
@@ -21,7 +26,6 @@ impl EventService {
         Ok(Self::row_to_event(row))
     }
 
-    #[allow(dead_code)] // Used by SSE streaming in sessions.rs through db directly
     pub async fn list(&self, session_id: Uuid, since_sequence: Option<i32>) -> Result<Vec<Event>> {
         let rows = self.db.list_events(session_id, since_sequence).await?;
         Ok(rows.into_iter().map(Self::row_to_event).collect())
